+++ FindPOLine/OrderLinesFilters.test.js
@@ -1,11 +1,14 @@
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import { render } from '@folio/jest-config-stripes/testing-library/react';
+import { render, act } from '@folio/jest-config-stripes/testing-library/react';
+import { AcqDateRangeFilter } from '@folio/stripes-acq-components';
 
 import { OrderLinesFilters } from './OrderLinesFilters';
 
-jest.mock('@folio/stripes-acq-components/lib/ExpenseClassFilter/useExpenseClassOptions', () => ({
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
   useExpenseClassOptions: jest.fn().mockReturnValue([]),
+  AcqDateRangeFilter: jest.fn().mockReturnValue('AcqDateRangeFilter'),
 }));
 
 const queryClient = new QueryClient();
@@ -17,11 +20,12 @@ const wrapper = ({ children }) => (
   </QueryClientProvider>
 );
 
-const renderOrderLinesFilters = () => render(
+const renderOrderLinesFilters = (props) => render(
   <OrderLinesFilters
     activeFilters={{}}
     onChange={jest.fn}
     applyFilters={jest.fn}
+    {...props}
   />,
   { wrapper },
 );
@@ -39,14 +43,26 @@ describe('OrderLinesFilters component', () => {
     expect(getByText('ui-orders.line.accordion.location')).toBeDefined();
     expect(getByText('stripes-acq-components.filter.fundCode')).toBeDefined();
     expect(getByText('ui-orders.poLine.orderFormat')).toBeDefined();
-    expect(getByText('ui-orders.poLine.dateCreated')).toBeDefined();
     expect(getByText('ui-orders.line.accordion.vendor')).toBeDefined();
     expect(getByText('ui-orders.filter.collection')).toBeDefined();
     expect(getByText('ui-orders.filter.rush')).toBeDefined();
     expect(getByText('ui-orders.eresource.accessProvider')).toBeDefined();
-    expect(getByText('ui-orders.eresource.expectedActivation')).toBeDefined();
     expect(getByText('stripes-acq-components.filter.expenseClass')).toBeDefined();
-    expect(getByText('ui-orders.export.exportDate')).toBeDefined();
     expect(getAllByText('ui-orders.line.accordion.donor')).toHaveLength(2);
+  });
+
+  it('should display filters', async () => {
+    AcqDateRangeFilter.mockClear();
+    const applyFilters = jest.fn();
+    const filterValues = { name: 'dateCreated', values: ['2021-01-01:2021-01-31'] };
+
+    renderOrderLinesFilters({
+      applyFilters,
+    });
+
+    act(() => {
+      AcqDateRangeFilter.mock.calls[0][0].onChange(filterValues);
+    });
+    expect(applyFilters).toHaveBeenCalledWith(filterValues.name, filterValues.values[0]);
   });
 });